Floor Plan: the canvas was drawing everything twice, and stairs now step by half a cell
Two things behind "the staircase feels sticky".

The canvas was rendering the staircases, the whole wall set and the wall
highlight TWICE. An earlier edit that meant to move the stairs group under the
walls copied it instead, and the bug was invisible because the duplicate paints
exactly over the original — same data, same geometry, pixel for pixel. Every
pointermove was hit-testing twice the SVG it needed to, and every wall was
being stroked twice. Caught by counting rects on the canvas: two for a plan
with one room and one staircase, where it had been three.

Stairs now position in half cells. A staircase owns no grid cells — it is drawn
over whatever room it sits in — so nothing forces it onto whole ones. Whole-cell
snapping meant the pointer had to travel half a cell, about 22px on screen,
before anything happened at all, and then it jumped a full cell: exactly the
feel being described. Measured after the change: a 20px drag now moves it
(2 -> 2.5) and an 8px nudge registers. Doors already worked this way. Size stays
in whole cells; only position is finer, which also lets a flight centre itself
in an odd-width space.

Worth recording for next time: a synthetic drag cannot judge this. Driving one
through the browser produced two pointermove events a full second apart, so it
proves what the code does but says nothing about how it feels. The duplicate
render was found by reading the file, not by measuring.

+++ lib/floor-plan/types.ts
@@ -164,8 +164,17 @@ export type RemovedWall = z.infer<typeof removedWallSchema>;
  */
 export const stairSchema = z.object({
   id: z.string().min(1),
-  x: z.number().int().min(0),
-  y: z.number().int().min(0),
+  /**
+   * Position is in HALF cells, size in whole ones.
+   *
+   * A staircase owns no cells — it is drawn over whatever room it sits in — so nothing forces
+   * it onto the grid the way a room is forced. Whole-cell positioning made it feel sticky to
+   * drag: half a cell is ~13px on screen, so the pointer moved that far before anything
+   * happened and then it jumped a whole cell. Half steps halve that, and let a flight centre
+   * itself in an odd-width space. Doors already do the same.
+   */
+  x: z.number().min(0),
+  y: z.number().min(0),
   w: z.number().int().min(1),
   h: z.number().int().min(1),
   dir: z.enum(["up", "down"]).default("up"),
